requirements.txt
yt-dlp
flet
python-telegram-bot
python-dotenv
speedtest-cli
telethon
cryptg
